fix bug and case issue

+++ src/client/src/tscUtil.c
@@ -4409,7 +4409,7 @@ bool tscHasReachLimitation(SQueryInfo* pQueryInfo, SSqlRes* pRes) {
 
 char* tscGetErrorMsgPayload(SSqlCmd* pCmd) { return pCmd->payload; }
 
-int32_t tscGetErrorMsgLength(SSqlCmd* pCmd) { return strlen(pCmd->payload); }
+int32_t tscGetErrorMsgLength(SSqlCmd* pCmd) { return (int32_t)strlen(pCmd->payload); }
 
 /**
  *  If current vnode query does not return results anymore (pRes->numOfRows == 0), try the next vnode if exists,
